[fix] moved aggr_test_schema to test_utils
This avoids that its duplicated in the benchmark package.
